+++ src/console/controllers/FieldsController.php
@@ -8,6 +8,7 @@ use Craft;
 use craft\console\Controller;
 use craft\db\Query;
 use craft\db\Table;
+use craft\helpers\Db;
 use craft\helpers\Queue;
 use craft\i18n\Translation;
 use craft\queue\jobs\ApplyNewPropagationMethod;
@@ -52,6 +53,24 @@ class FieldsController extends Controller
         }
 
         return $options;
+    }
+
+    /**
+     * Changes `null` block structure site IDs to the primary site ID.
+     *
+     * @return int
+     * @since 4.1.0
+     */
+    public function actionFixBlockStructureSiteIds(): int
+    {
+        Db::update('{{%neoblockstructures}}', [
+            'siteId' => Craft::$app->getSites()->getPrimarySite()->id,
+        ], [
+            'siteId' => null,
+        ]);
+        $this->stdout('Done.' . PHP_EOL);
+
+        return ExitCode::OK;
     }
 
     /**