+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added `benf\neo\console\controllers\BlockTypesController::actionFixFieldLayouts()` (`php craft neo/block-types/fix-field-layouts` console command)
+- Added `benf\neo\console\controllers\FieldsController::actionFixBlockStructureSiteIds()` (`php craft neo/fields/fix-block-structure-site-ids` console command)
 
 ### Fixed
 - Fixed an error that could occur when changing a Neo field's propagation method, if a Neo block structure for that field exists for an owner element that no longer has the field on its field layout